+++ bwapi/include/BWAPI/Client/GameImpl.h
@@ -111,7 +111,7 @@ namespace BWAPI
       virtual std::set<Unit*>& getUnitsOnTile(int x, int y);
       virtual std::set<Unit*>  getUnitsInRectangle(int left, int top, int right, int bottom);
       virtual std::set<Unit*>  getUnitsInRectangle(BWAPI::Position topLeft, BWAPI::Position bottomRight);
-      virtual std::set<Unit*>  getUnitsInRadius(BWAPI::Position center, double radius);
+      virtual std::set<Unit*>  getUnitsInRadius(BWAPI::Position center, int radius);
       virtual Error getLastError() const;
       virtual bool  setLastError(BWAPI::Error e);
